food-story-app/requirements.txt
django>=4.2
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
anthropic>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
